+++ src/comet/emulator/__init__.py
@@ -1,5 +1,16 @@
-from .emulator import Emulator, message, run
+from .emulator import Emulator, message
+from .response import TextResponse, BinaryResponse, RawResponse
 from .iec60488 import IEC60488Emulator
 from .resource import open_emulator
+from .tcpserver import run
 
-__all__ = ["Emulator", "message", "run", "IEC60488Emulator", "open_emulator"]
+__all__ = [
+    "Emulator",
+    "message",
+    "TextResponse",
+    "BinaryResponse",
+    "RawResponse",
+    "IEC60488Emulator",
+    "open_emulator",
+    "run",
+]