+++ tests/test_emulator_nkt_photonics_pilas.py
@@ -34,11 +34,11 @@ def test_tune_mode(emulator):
 
 def test_tune(emulator):
     assert emulator("tune?") == "tune value:\t\t     0.00 %"
-    assert emulator("tune=500") == None
+    assert emulator("tune=500") == "done"
     assert emulator("tune?") == "tune value:\t\t     50.00 %"
-    assert emulator("tune=1000") == None
+    assert emulator("tune=1000") == "done"
     assert emulator("tune?") == "tune value:\t\t     100.00 %"
-    assert emulator("tune=0") == None
+    assert emulator("tune=0") == "done"
     assert emulator("tune?") == "tune value:\t\t     0.00 %"
 
 
@@ -53,5 +53,10 @@ def test_frequency(emulator):
 def test_laser_diode_temperature(emulator):
     assert emulator("ldtemp?") == "LD temp.:\t\tgood"
     emulator.laser_diode_temperature = False
-
     assert emulator("ldtemp?") == "LD temp.:\t\tbad"
+
+
+def test_laser_head_temperature(emulator):
+    assert emulator("lht?") == "laser head temp.:\t     25.0 °C".encode("latin-1")
+    emulator.laser_head_temperature += 1.0
+    assert emulator("lht?") == "laser head temp.:\t     26.0 °C".encode("latin-1")